Do not manually clear the RX flag from MCP2515_REG_CANINTF.
It seems that the MCP2515's Read RX buffer instruction clears this flag
automatically (see section 12.4, p.63 in the datasheet).  Clearing it a
second time is not only redundant, but dangerous.  If there is a flood
of incoming CAN messages, clearing of the RX flag a second time can
cause a message loss.